+++ tests/testthat/test-e_to_x.R
@@ -1,0 +1,8 @@
+test_that("loading the phi constants table works as expected", {
+  phi_table <- load_phi_constants_table()
+  colnames <- phi_table %>%
+    names()
+  expect_equal(colnames, c("Product", "phi", "is.useful"))
+})
+
+